Improved pid package
