+++ pid/pid_darwin.go
@@ -1,0 +1,13 @@
+package pid
+
+// ////////////////////////////////////////////////////////////////////////////////// //
+//                                                                                    //
+//                     Copyright (c) 2009-2016 Essential Kaos                         //
+//      Essential Kaos Open Source License <http://essentialkaos.com/ekol?en>         //
+//                                                                                    //
+// ////////////////////////////////////////////////////////////////////////////////// //
+
+// IsWorks return if process pid is not -1
+func IsWorks(name string) bool {
+	return Get(name) != -1
+}
