+++ pid/pid_linux.go
@@ -1,0 +1,27 @@
+package pid
+
+// ////////////////////////////////////////////////////////////////////////////////// //
+//                                                                                    //
+//                     Copyright (c) 2009-2016 Essential Kaos                         //
+//      Essential Kaos Open Source License <http://essentialkaos.com/ekol?en>         //
+//                                                                                    //
+// ////////////////////////////////////////////////////////////////////////////////// //
+
+import (
+	"fmt"
+
+	"pkg.re/essentialkaos/ek.v2/fsutil"
+)
+
+// ////////////////////////////////////////////////////////////////////////////////// //
+
+// IsWorks return if process with pid from pid file is works
+func IsWorks(name string) bool {
+	pid := Get(name)
+
+	if pid == -1 {
+		return false
+	}
+
+	return fsutil.IsExist(fmt.Sprintf("/proc/%d", pid))
+}
